[Agent] extract component builder util
